IGNITE-26601 / deduplicate more code
