+++ modules/network/src/test/java/org/apache/ignite/internal/network/recovery/RecoveryAcceptorHandshakeManagerTest.java
@@ -45,14 +45,11 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.BooleanSupplier;
-import org.apache.ignite.internal.configuration.testframework.ConfigurationExtension;
 import org.apache.ignite.internal.configuration.testframework.InjectConfiguration;
 import org.apache.ignite.internal.lang.NodeStoppingException;
 import org.apache.ignite.internal.network.ClusterNodeImpl;
 import org.apache.ignite.internal.network.ConstantClusterIdSupplier;
-import org.apache.ignite.internal.network.NetworkMessagesFactory;
 import org.apache.ignite.internal.network.OutNetworkObject;
-import org.apache.ignite.internal.network.RecipientLeftException;
 import org.apache.ignite.internal.network.configuration.AckConfiguration;
 import org.apache.ignite.internal.network.handshake.HandshakeException;
 import org.apache.ignite.internal.network.handshake.NoOpHandshakeEventLoopSwitcher;
@@ -61,20 +58,16 @@ import org.apache.ignite.internal.network.netty.NettySender;
 import org.apache.ignite.internal.network.recovery.message.HandshakeRejectedMessage;
 import org.apache.ignite.internal.network.recovery.message.HandshakeRejectionReason;
 import org.apache.ignite.internal.network.recovery.message.HandshakeStartResponseMessage;
-import org.apache.ignite.internal.testframework.BaseIgniteAbstractTest;
 import org.apache.ignite.internal.version.DefaultIgniteProductVersionSource;
 import org.apache.ignite.network.NetworkAddress;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
 import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
 
-@ExtendWith({MockitoExtension.class, ConfigurationExtension.class})
-class RecoveryAcceptorHandshakeManagerTest extends BaseIgniteAbstractTest {
+class RecoveryAcceptorHandshakeManagerTest extends HandshakeManagerTest {
     private static final UUID LOWER_ID = new UUID(1, 1);
     private static final UUID HIGHER_ID = new UUID(2, 2);
 
@@ -87,8 +80,6 @@ class RecoveryAcceptorHandshakeManagerTest extends BaseIgniteAbstractTest {
     private static final String INITIATOR_HOST = "initiator-host";
 
     private static final int PORT = 1000;
-
-    private static final NetworkMessagesFactory MESSAGE_FACTORY = new NetworkMessagesFactory();
 
     private static final UUID CORRECT_CLUSTER_ID = new UUID(11, 12);
 
@@ -260,19 +251,6 @@ class RecoveryAcceptorHandshakeManagerTest extends BaseIgniteAbstractTest {
     void gettingHandshakeRejectedMessageWithReasonStoppingCausesHandshakeToBeFinishedWithRecipientLeftException() {
         RecoveryAcceptorHandshakeManager manager = acceptorHandshakeManager(LOWER_ID);
 
-        CompletableFuture<NettySender> localHandshakeFuture = manager.localHandshakeFuture();
-        CompletionStage<NettySender> finalHandshakeFuture = manager.finalHandshakeFuture();
-
-        manager.onMessage(handshakeRejectedMessageWithReason(HandshakeRejectionReason.STOPPING));
-
-        assertWillThrowFast(localHandshakeFuture, RecipientLeftException.class);
-        assertWillThrowFast(finalHandshakeFuture.toCompletableFuture(), RecipientLeftException.class);
-    }
-
-    private static HandshakeRejectedMessage handshakeRejectedMessageWithReason(HandshakeRejectionReason reason) {
-        return MESSAGE_FACTORY.handshakeRejectedMessage()
-                .message("Rejected")
-                .reasonString(reason.toString())
-                .build();
+        assertThatRejectionWithStoppingCausesRecipientLeftException(manager);
     }
 }
