+++ modules/network/src/test/java/org/apache/ignite/internal/network/recovery/RecoveryInitiatorHandshakeManagerTest.java
@@ -46,14 +46,12 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.BooleanSupplier;
-import org.apache.ignite.internal.configuration.testframework.ConfigurationExtension;
 import org.apache.ignite.internal.configuration.testframework.InjectConfiguration;
 import org.apache.ignite.internal.lang.NodeStoppingException;
 import org.apache.ignite.internal.network.ClusterNodeImpl;
 import org.apache.ignite.internal.network.ConstantClusterIdSupplier;
 import org.apache.ignite.internal.network.NetworkMessagesFactory;
 import org.apache.ignite.internal.network.OutNetworkObject;
-import org.apache.ignite.internal.network.RecipientLeftException;
 import org.apache.ignite.internal.network.configuration.AckConfiguration;
 import org.apache.ignite.internal.network.handshake.ChannelAlreadyExistsException;
 import org.apache.ignite.internal.network.handshake.HandshakeException;
@@ -64,23 +62,19 @@ import org.apache.ignite.internal.network.recovery.message.HandshakeRejectedMess
 import org.apache.ignite.internal.network.recovery.message.HandshakeRejectionReason;
 import org.apache.ignite.internal.network.recovery.message.HandshakeStartMessage;
 import org.apache.ignite.internal.properties.IgniteProductVersion;
-import org.apache.ignite.internal.testframework.BaseIgniteAbstractTest;
 import org.apache.ignite.internal.version.DefaultIgniteProductVersionSource;
 import org.apache.ignite.network.NetworkAddress;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
 import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
 
-@ExtendWith({MockitoExtension.class, ConfigurationExtension.class})
 @Timeout(10)
-class RecoveryInitiatorHandshakeManagerTest extends BaseIgniteAbstractTest {
+class RecoveryInitiatorHandshakeManagerTest extends HandshakeManagerTest {
     private static final UUID LOWER_ID = new UUID(1, 1);
     private static final UUID HIGHER_ID = new UUID(2, 2);
 
@@ -340,19 +334,6 @@ class RecoveryInitiatorHandshakeManagerTest extends BaseIgniteAbstractTest {
     void gettingHandshakeRejectedMessageWithReasonStoppingCausesHandshakeToBeFinishedWithRecipientLeftException() {
         RecoveryInitiatorHandshakeManager manager = initiatorHandshakeManager(LOWER_ID);
 
-        CompletableFuture<NettySender> localHandshakeFuture = manager.localHandshakeFuture();
-        CompletionStage<NettySender> finalHandshakeFuture = manager.finalHandshakeFuture();
-
-        manager.onMessage(handshakeRejectedMessageWithReason(HandshakeRejectionReason.STOPPING));
-
-        assertWillThrowFast(localHandshakeFuture, RecipientLeftException.class);
-        assertWillThrowFast(finalHandshakeFuture.toCompletableFuture(), RecipientLeftException.class);
-    }
-
-    private static HandshakeRejectedMessage handshakeRejectedMessageWithReason(HandshakeRejectionReason reason) {
-        return MESSAGE_FACTORY.handshakeRejectedMessage()
-                .message("Rejected")
-                .reasonString(reason.toString())
-                .build();
+        assertThatRejectionWithStoppingCausesRecipientLeftException(manager);
     }
 }
